requirements.txt
pythainlp
sklearn-crfsuite
python-docx
